fix: プレビュー画像を表示する新メソッドの追加
